Big ugly commit
This may or may not be functional code; I'm committing it so I have some kind of baseline. I'm fairly certain that this code worked 15 minutes ago, then I undid and redid a couple things and must ahve broken something. Now jQuery has mysteriously quit ($ is not a function, .val is not a function).

Future commits on this branched will be untested, unless I figure out how to convince jQuery to give me another chance.